olap-ui/be/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
psycopg2-binary==2.9.10
pydantic==2.11.1
